+++ config/additional-settings.dist.php
@@ -20,11 +20,22 @@ return [
             'secret' => '',
         ],
         
+        // Google QR Code title
+        '2fa_qrc_title' => null,
+        
+        // User validation
+        'validation' => [
+            'min_user_name_length'    => 4,
+            'max_user_name_length'    => 50,
+            'min_password_length'     => 6,
+            'allowed_user_name_chars' => str_split('abcdefghijklmnopqrstuvwxyz0123456789-_'),
+        ],
+        
         // Locale settings
         'locale' => [
-            'process' => \App\Utility\LanguageUtility::LOCALE_URL | \App\Utility\LanguageUtility::DOMAIN_DISABLED,
+            'process'     => \App\Utility\LanguageUtility::LOCALE_URL | \App\Utility\LanguageUtility::DOMAIN_DISABLED,
             'auto_detect' => TRUE,
-            'code' => 'en-US', // default / current language
+            'code'        => 'en-US', // default / current language
             'active' => [
                 'en-US' => 'imhh-fs.localhost',
                 'de-DE' => 'de.imhh-fs.localhost',